+++ zanata-war/src/main/webapp-jboss/WEB-INF/web.xml
@@ -148,7 +148,9 @@
   </listener>
 
   <listener>
-    <listener-class>org.zanata.servlet.ZanataDatabaseMetadataServletListener</listener-class>
+    <!-- this listener needs to appear before liquibase listener to log useful
+    database metadata for troubleshooting potential errors-->
+    <listener-class>org.zanata.database.DatabaseMetadataListener</listener-class>
   </listener>
 
   <listener>
